Share fetch code between client/server -> big-brain
Since both are calling the same getQuota endpoint - share the
code so it's easier to not mess it up when changing it.
