+++ app/lib/convexUsage.ts
@@ -1,7 +1,21 @@
-const VITE_PROVISION_HOST = import.meta.env.VITE_PROVISION_HOST || 'https://api.convex.dev';
+export type CheckTokenUsageResponse =
+  | {
+      status: 'success';
+      tokensUsed: number;
+      tokensQuota: number;
+    }
+  | {
+      status: 'error';
+      httpStatus: number;
+      httpBody: string;
+    };
 
-export async function getTokenUsage(convexAuthToken: string, teamSlug: string) {
-  const url = `${VITE_PROVISION_HOST}/api/dashboard/teams/${teamSlug}/usage/get_token_info`;
+export async function getTokenUsage(
+  provisionHost: string,
+  convexAuthToken: string,
+  teamSlug: string,
+): Promise<CheckTokenUsageResponse> {
+  const url = `${provisionHost}/api/dashboard/teams/${teamSlug}/usage/get_token_info`;
   const response = await fetch(url, {
     method: 'GET',
     headers: {
@@ -13,6 +27,14 @@ export async function getTokenUsage(convexAuthToken: string, teamSlug: string) {
     const body = await response.text();
     throw new Error(`Failed to fetch usage: ${response.statusText}: ${body}`);
   }
-  const { tokensUsed, tokensQuota } = await response.json();
-  return { tokensUsed, tokensQuota };
+  if (!response.ok) {
+    const body = await response.text();
+    return {
+      status: 'error',
+      httpStatus: response.status,
+      httpBody: body,
+    };
+  }
+  const { tokensUsed, tokensQuota }: { tokensUsed: number; tokensQuota: number } = await response.json();
+  return { status: 'success', tokensUsed, tokensQuota };
 }
